+++ backend/src/main/java/com/huy/husci/repository/AnnouncementRepository.java
@@ -1,7 +1,8 @@
 package com.huy.husci.repository;
 
-import com.huy.husci.model.Announcement;
 import org.springframework.data.mongodb.repository.MongoRepository;
+
+import com.huy.husci.repository.entity.Announcement;
 
 public interface AnnouncementRepository extends MongoRepository<Announcement, String> {
 }
