+++ backend/src/main/java/com/huy/husci/utils/event/AnnouncementCreatedEvent.java
@@ -1,6 +1,6 @@
-package com.huy.husci.event;
+package com.huy.husci.utils.event;
 
-import com.huy.husci.model.Announcement;
+import com.huy.husci.repository.entity.Announcement;
 
 public class AnnouncementCreatedEvent {
     private final Announcement announcement;
