+++ backend/src/main/java/com/huy/husci/utils/event/StudentDeletedEvent.java
@@ -1,6 +1,6 @@
-package com.huy.husci.event;
+package com.huy.husci.utils.event;
 
-import com.huy.husci.model.Student;
+import com.huy.husci.repository.entity.Student;
 
 public class StudentDeletedEvent {
     private final Student student;
